Remove session id
Was added for some reason in https://github.com/SuffolkLITLab/docassemble-InterviewStats/commit/33234abd2112dcf73ebd594b1530ad79bd241a05,
my only guess as to why is that we needed to display the user defined tags,
and docassemble internally calls that a key. So likely the first
implementation tried to get the key from the JSON table, but it was the wrong
value.

Should be able to remove it without harm, but TBH it needs an upstream patch,
or we should start using a different value.